pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Azim Akmatov <azim.akmatov1@gmail.com>"]
description = ""
name = "bseyes"
version = "0.1.0"
[tool.poetry.dependencies]
django = "^3.0"
python = "^3.7"
djangorestframework = "^3.11"
[tool.poetry.dev-dependencies]
